+++ backend-lib/src/main/java/cash/z/ecc/android/sdk/internal/model/TorClient.kt
@@ -4,22 +4,24 @@ import cash.z.ecc.android.sdk.internal.ext.existsSuspend
 import cash.z.ecc.android.sdk.internal.ext.mkdirsSuspend
 import cash.z.ecc.android.sdk.internal.jni.RustBackend
 import dalvik.annotation.optimization.CriticalNative
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
 import java.io.File
 import java.math.BigDecimal
 
 class TorClient(
     private val nativeHandle: Long?,
 ) {
-    fun dispose() {
+    suspend fun dispose() = withContext(Dispatchers.IO) {
         nativeHandle?.let { freeTorRuntime(it) }
     }
 
-    fun getExchangeRateUsd(): BigDecimal {
-        return getExchangeRateUsd(nativeHandle!!)
+    suspend fun getExchangeRateUsd(): BigDecimal = withContext(Dispatchers.IO) {
+        getExchangeRateUsd(nativeHandle!!)
     }
 
     companion object {
-        suspend fun new(torDir: File): TorClient {
+        suspend fun new(torDir: File): TorClient = withContext(Dispatchers.IO) {
             RustBackend.loadLibrary()
 
             // Ensure that the directory exists.
@@ -28,7 +30,7 @@ class TorClient(
                 error("${torDir.path} directory does not exist and could not be created.")
             }
 
-            return TorClient(createTorRuntime(torDir.path))
+            TorClient(createTorRuntime(torDir.path))
         }
 
         //