+++ sdk-incubator-lib/src/androidTest/kotlin/cash/z/ecc/android/sdk/fixture/FiatCurrencyFixture.kt
@@ -3,7 +3,5 @@ package cash.z.ecc.android.sdk.fixture
 import cash.z.ecc.android.sdk.model.FiatCurrency
 
 object FiatCurrencyFixture {
-    const val USD = "USD"
-
-    fun new(code: String = USD) = FiatCurrency(code)
+    fun new() = FiatCurrency.USD
 }