+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/model/FiatCurrencyConversion.kt
@@ -10,7 +10,7 @@ import kotlinx.datetime.Instant
  * the server so it shouldn't have issues with client-side clock inaccuracy.
  * @param priceOfZec The conversion rate of ZEC to the fiat currency.
  */
-data class CurrencyConversion(
+data class FiatCurrencyConversion(
     val fiatCurrency: FiatCurrency,
     val timestamp: Instant,
     val priceOfZec: Double
@@ -19,17 +19,6 @@ data class CurrencyConversion(
         require(priceOfZec > 0) { "priceOfZec must be greater than 0" }
         require(priceOfZec.isFinite()) { "priceOfZec must be finite" }
     }
-}
 
-/**
- * Represents an ISO 4217 currency code.
- */
-@Suppress("MagicNumber")
-data class FiatCurrency(val code: String) {
-    init {
-        require(code.length == 3) { "Fiat currency code must be 3 characters long." }
-
-        // TODO [#532]: Add another check to make sure the code is in the known ISO currency code list.
-        // TODO [#532]: https://github.com/zcash/secant-android-wallet/issues/532
-    }
+    companion object
 }